fix(migration): 修正數字資料型別 & 增加 avatar_pic 欄位

+++ database/migrations/2024_08_27_035257_create_dailies_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained();;
             $table->date('date')->unique();
-            $table->decimal('weight', 4, 1)->nullable();
-            $table->decimal('body_fat', 4, 3)->nullable();
+            $table->float('weight', 4, 1)->nullable();
+            $table->float('body_fat', 4, 3)->nullable();
             $table->string('note', 1000)->nullable();
             $table->integer('water_morning')->nullable();
             $table->integer('water_afternoon')->nullable();
